refactor: Update df_to_bigwig function to accept a column parameter
The df_to_bigwig function in utils.py has been updated to accept a column parameter. This allows the user to specify which column from the DataFrame should be used as the values for the bigwig file. The function now uses the specified column when adding entries to the bigwig file.

This change improves the flexibility and usability of the df_to_bigwig function by allowing the user to choose the desired column for the bigwig file, instead of being limited to a fixed column name.