Refactor StridedFit to include a terminal fit condition
This commit refactors the StridedFit class in the fit.py file to include a terminal fit condition. The terminal fit condition is based on whether a fit has reached a low RMSEA value and is not successful. This change improves the termination criteria for the fitting process.

Also, update the default value of rmsea_tr in ProcessorConfig

This commit updates the default value of rmsea_tr in the ProcessorConfig class in the models.py file from 0.0 to 0.05. This change ensures that the RMSEA threshold is set to a more appropriate value.